Updated README to include instructions on opening project in VS.
Removed the FK constraints on Employee as it was IMPOSSIBLE to add anything to an empty self referencing table.

Creating a new user now works, although checking if EMPID does not work intended. Creates an aspnet_user but does not create an Employee, need to fix empid check logic BEFORE aspnet_user is created.